:sparkles: Namespaces: Moving out project state management
